loader,workspace: safe FFI wrappers; document the unsafe surface
Audit of the loader found 393 unsafe blocks and none that could simply be
deleted: every one serves an FFI or low-level requirement. So this pass does
two things — make each remaining unsafe operation explicit and documented, and
wrap the repeated shapes once so call sites stop re-deriving their invariants.

- `jni_utils` (new): `JniStringGuard` pairs GetStringUTFChars with
  ReleaseStringUTFChars in a `Drop`, and `cstr_to_owned` copies a nullable C
  string immediately — the borrow never outlives the pointer's storage, so the
  helper is genuinely safe where a returned reference could not be. 17
  `CStr::from_ptr` blocks and both specialize paths
  (app_specialize/native_specialize) now go through them.
- `ModuleSnapshot::{get,get_mut,iter,iter_mut}`: bounds-checked accessors
  replace `base.add(i)` arithmetic in load_modules and the cleanup loop.
- Call sites that previously trusted a non-NULL pointer (hook-list replay,
  `dlerror`) take the documented well-defined path — empty string / "(null)" —
  instead of UB.
- Stale `hook.c <line-range>` references are gone (209 lines): RS owns this
  behavior, and C line numbers no longer describe the tree being read.
- Dead code that only existed for C parity is removed: ReZygiskInfo,
  rezygiskd_get_info, free_rezygisk_info, rezygiskd_zygote_restart.
- CHANGELOG: v1.0.1-rs entry for the pass.

Device evidence for exactly this source state: one generation staged for both
ABIs (RZGEN-1.0.0+g7e151c4-dirty.w186228b5), 19/19 stage checks, 24/24
post-boot assertions with both loaders reporting the new generation, and a
45-minute soak PASS (diag/20260923_091923_post-plan-cstr).

+++ crates/loader/src/hook_register.rs
@@ -28,7 +28,7 @@ macro_rules! dloge {
     ($($arg:tt)*) => {{ rz_common::loge!(crate::hook_register::TAG, $($arg)*); }};
 }
 
-/// hook.c `hook_register` (1291-1301): forward one PLT hook to PLTI.
+/// hook.c `hook_register`: forward one PLT hook to PLTI.
 ///
 /// `backup` is the address of the old_* static (NULL when the caller does
 /// not keep a backup); plti writes it only while the slot is still 0,
@@ -63,7 +63,7 @@ pub unsafe fn hook_register(
     true
 }
 
-/// hook.c `hook_unregister` (1303-1313): remove one PLT hook from PLTI.
+/// hook.c `hook_unregister`: remove one PLT hook from PLTI.
 ///
 /// The C `plti_remove_hook` reads the original callback out of `*backup`
 /// and rejects NULL (logged by the plti crate); a NULL `backup` slot
@@ -97,7 +97,7 @@ pub unsafe fn hook_unregister(
     true
 }
 
-/// hook.c `hook_functions` (1327-1336): init PLTI, add libandroid_runtime.so
+/// hook.c `hook_functions`: init PLTI, add libandroid_runtime.so
 /// and register the four core loader hooks.
 pub fn hook_functions() -> u32 {
     // Status bitmask — 0 = all four hooks registered.
@@ -119,7 +119,7 @@ pub fn hook_functions() -> u32 {
             "libandroid_runtime.so",
             "fork",
             false,
-            crate::fork_hooks::fork as usize as *mut c_void,
+            crate::fork_hooks::fork as *mut c_void,
             crate::fork_hooks::OLD_FORK.as_ptr() as *mut *mut c_void,
         ) {
             status |= 0x2;
@@ -128,7 +128,7 @@ pub fn hook_functions() -> u32 {
             "libandroid_runtime.so",
             "strdup",
             false,
-            crate::fork_hooks::strdup as usize as *mut c_void,
+            crate::fork_hooks::strdup as *mut c_void,
             crate::fork_hooks::OLD_STRDUP.as_ptr() as *mut *mut c_void,
         ) {
             status |= 0x4;
@@ -137,7 +137,7 @@ pub fn hook_functions() -> u32 {
             "libandroid_runtime.so",
             "property_get",
             false,
-            crate::fork_hooks::property_get as usize as *mut c_void,
+            crate::fork_hooks::property_get as *mut c_void,
             crate::fork_hooks::OLD_PROPERTY_GET.as_ptr() as *mut *mut c_void,
         ) {
             status |= 0x8;
@@ -146,7 +146,7 @@ pub fn hook_functions() -> u32 {
             "libandroid_runtime.so",
             "_ZNK18FileDescriptorInfo14ReopenOrDetach",
             true,
-            crate::fork_hooks::_ZNK18FileDescriptorInfo14ReopenOrDetach as usize as *mut c_void,
+            crate::fork_hooks::_ZNK18FileDescriptorInfo14ReopenOrDetach as *mut c_void,
             crate::fork_hooks::OLD__ZNK18FileDescriptorInfo14ReopenOrDetach.as_ptr() as *mut *mut c_void,
         ) {
             status |= 0x10;
@@ -156,7 +156,7 @@ pub fn hook_functions() -> u32 {
     status
 }
 
-/// hook.c `hook_unloader` (1338-1355): once libart.so is mapped, hook
+/// hook.c `hook_unloader`: once libart.so is mapped, hook
 /// pthread_attr_setstacksize, drop the property_get hook that got us here,
 /// then load modules early (before the system server fork) to spread them
 /// through all Zygotes.
@@ -172,7 +172,7 @@ pub fn hook_unloader() {
             "libart.so",
             "pthread_attr_setstacksize",
             false,
-            crate::fork_hooks::pthread_attr_setstacksize as usize as *mut c_void,
+            crate::fork_hooks::pthread_attr_setstacksize as *mut c_void,
             crate::fork_hooks::OLD_PTHREAD_ATTR_SETSTACKSIZE.as_ptr() as *mut *mut c_void,
         );
     }
